requirements.txt
asn1crypto==0.24.0
botocore==1.12.29
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.1
cryptography==2.3.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
idna==2.7
imageio==2.4.1
kiwisolver==1.0.1
matplotlib==3.0.0
mkl-fft==1.0.6
mkl-random==1.0.1
networkx==2.2
numpy==1.15.3
olefile==0.46
Pillow==5.3.0
pycosat==0.6.3
pycparser==2.18
pyOpenSSL==18.0.0
pyparsing==2.2.2
pypng==0.0.18
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
requests==2.19.1
ruamel-yaml==0.15.46
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
tornado==5.1.1
urllib3==1.23
